adapt to the changes from master
integration-tests -> tests
