--- integration-tests/src/test/java/org/eclipse/dirigible/integration/tests/ui/Dirigible.java
@@ -1,76 +1,0 @@
-/*
- * Copyright (c) 2023 SAP SE or an SAP affiliate company and Eclipse Dirigible contributors
- *
- * All rights reserved. This program and the accompanying materials are made available under the
- * terms of the Eclipse Public License v2.0 which accompanies this distribution, and is available at
- * http://www.eclipse.org/legal/epl-v20.html
- *
- * SPDX-FileCopyrightText: 2023 SAP SE or an SAP affiliate company and Eclipse Dirigible
- * contributors SPDX-License-Identifier: EPL-2.0
- */
-package org.eclipse.dirigible.integration.tests.ui;
-
-import org.eclipse.dirigible.integration.tests.ui.framework.Browser;
-import org.eclipse.dirigible.integration.tests.ui.framework.HtmlAttribute;
-import org.eclipse.dirigible.integration.tests.ui.framework.HtmlElementType;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import com.codeborne.selenide.SelenideElement;
-
-public class Dirigible {
-
-    private static final Logger LOGGER = LoggerFactory.getLogger(Dirigible.class);
-
-    private static final String LOGIN_PAGE_TITLE = "Please sign in";
-
-    private static final String ROOT_PATH = "/";
-
-    private static final String USERNAME = "admin";
-    private static final String PASSWORD = "admin";
-
-    private static final String USERNAME_FIELD_ID = "username";
-    private static final String PASSWORD_FIELD_ID = "password";
-    private static final String SUBMIT_TYPE = "submit";
-
-    private static final String SIGN_IN_BUTTON_TEXT = "Sign in";
-
-    private static final String HELP_MENU_BUTTON = "Help";
-
-    private static final String ADMIN_PANEL_OPTION = "Admin Panel";
-
-    private final Browser browser;
-
-    public Dirigible(Browser browser) {
-        this.browser = browser;
-    }
-
-    public void openHomePage() {
-        browser.openPath(ROOT_PATH);
-        login();
-    }
-
-    private void login() {
-        SelenideElement title = browser.waitUntilExist(HtmlElementType.TITLE);
-        String pageTitle = title.getOwnText();
-        if (!LOGIN_PAGE_TITLE.equals(pageTitle)) {
-            LOGGER.info("Skipping login");
-            return;
-        }
-        LOGGER.info("Logging...");
-        browser.enterTextInElementByAttributePattern(HtmlElementType.INPUT, HtmlAttribute.ID, USERNAME_FIELD_ID, USERNAME);
-        browser.enterTextInElementByAttributePattern(HtmlElementType.INPUT, HtmlAttribute.ID, PASSWORD_FIELD_ID, PASSWORD);
-        browser.clickElementByAttributePatternAndText(HtmlElementType.BUTTON, HtmlAttribute.TYPE, SUBMIT_TYPE, SIGN_IN_BUTTON_TEXT);
-    }
-
-    public void navigateToAdminPanel() {
-        openHelpMenu();
-        browser.clickElementByAttributePatternAndText(HtmlElementType.LI, HtmlAttribute.TITLE, ADMIN_PANEL_OPTION, ADMIN_PANEL_OPTION);
-    }
-
-    public void openHelpMenu() {
-        browser.clickElementByAttributePatternAndText(HtmlElementType.BUTTON, HtmlAttribute.IS_MENU, Boolean.TRUE.toString(),
-                HELP_MENU_BUTTON);
-    }
-
-
-}
