--- integration-tests/src/test/java/org/eclipse/dirigible/integration/tests/ui/framework/HtmlAttribute.java
@@ -1,28 +1,0 @@
-/*
- * Copyright (c) 2023 SAP SE or an SAP affiliate company and Eclipse Dirigible contributors
- *
- * All rights reserved. This program and the accompanying materials are made available under the
- * terms of the Eclipse Public License v2.0 which accompanies this distribution, and is available at
- * http://www.eclipse.org/legal/epl-v20.html
- *
- * SPDX-FileCopyrightText: 2023 SAP SE or an SAP affiliate company and Eclipse Dirigible
- * contributors SPDX-License-Identifier: EPL-2.0
- */
-package org.eclipse.dirigible.integration.tests.ui.framework;
-
-public enum HtmlAttribute {
-    ID("id"), //
-    TYPE("type"), //
-    IS_MENU("is-menu"), //
-    TITLE("title");
-
-    private final String attribute;
-
-    HtmlAttribute(String attribute) {
-        this.attribute = attribute;
-    }
-
-    public String getAttribute() {
-        return attribute;
-    }
-}
