+++ tests/pom.xml
@@ -43,6 +43,12 @@
 			<version>${selenide.version}</version>
 			<scope>test</scope>
 		</dependency>
+		<dependency>
+		    <groupId>org.junit.jupiter</groupId>
+		    <artifactId>junit-jupiter-params</artifactId>
+		    <version>5.8.2</version>
+		    <scope>test</scope>
+		</dependency>
 	</dependencies>
 
 	<dependencyManagement>
