+++ tests/src/test/java/org/eclipse/dirigible/integration/tests/api/SecurityIT.java
