+++ tests/src/test/java/org/eclipse/dirigible/integration/tests/ui/Dirigible.java
@@ -34,6 +34,10 @@ public class Dirigible {
 
     private static final String SIGN_IN_BUTTON_TEXT = "Sign in";
 
+    private static final String HELP_MENU_BUTTON = "Help";
+
+    private static final String ADMIN_PANEL_OPTION = "Admin Panel";
+
     private final Browser browser;
 
     public Dirigible(Browser browser) {
@@ -57,4 +61,16 @@ public class Dirigible {
         browser.enterTextInElementByAttributePattern(HtmlElementType.INPUT, HtmlAttribute.ID, PASSWORD_FIELD_ID, PASSWORD);
         browser.clickElementByAttributePatternAndText(HtmlElementType.BUTTON, HtmlAttribute.TYPE, SUBMIT_TYPE, SIGN_IN_BUTTON_TEXT);
     }
+
+    public void navigateToAdminPanel() {
+        openHelpMenu();
+        browser.clickElementByAttributePatternAndText(HtmlElementType.LI, HtmlAttribute.TITLE, ADMIN_PANEL_OPTION, ADMIN_PANEL_OPTION);
+    }
+
+    public void openHelpMenu() {
+        browser.clickElementByAttributePatternAndText(HtmlElementType.BUTTON, HtmlAttribute.IS_MENU, Boolean.TRUE.toString(),
+                HELP_MENU_BUTTON);
+    }
+
+
 }
