+++ tests/src/test/java/org/eclipse/dirigible/integration/tests/ui/framework/Browser.java
@@ -26,4 +26,6 @@ public interface Browser {
 
     SelenideElement waitUntilExist(HtmlElementType elementType);
 
+    void switchToWindow(String windowTitle);
+
 }
