+++ tests/src/test/java/org/eclipse/dirigible/integration/tests/ui/tests/SpringBootAdminIT.java
